move files around and organise correctly
